multiple images, animal type

+++ Aston-Animal-Society/app/Http/Controllers/AnimalController.php
@@ -10,16 +10,39 @@ use Gate;
 class AnimalController extends Controller
 {
 
+    public function index(){
+
+        $animals = Animal::select('*')->orderBy('name')->get();
+        $animals = $animals->filter(function($item){
+            if($item['availability'] == 'available'){
+                return $item;
+            }
+        }
+        )->toArray();
+
+        $sortingDetails = ['sort_by' => 'name', 'sorting_order' => 'asc'];
+
+        return view('animals.index', compact('animals', 'sortingDetails'));
+    }
+
+    public function sortby(Request $request){
+
+        $animals = Animal::select('*')->orderBy($request->input('sort_by'), $request->input('sorting_order'))->get();
+        $animals = $animals->filter(function($item){
+            if($item['availability'] == 'available'){
+                return $item;
+            }
+        }
+        )->toArray();
+        $sortingDetails = ['sort_by' => $request->input('sort_by'), 'sorting_order' => $request->input('sorting_order')];
+        return view('animals.index', compact('animals', 'sortingDetails'));
+    }
+
     public function show($id){
         $animal = Animal::find($id);
-        //return view('/show', array('animal' => $animal));
         return view('animals.show', compact('animal'));
     }
-/*
-    public function list(){
-        return view('/list', array('account'=>Animal::all()));
-    }
-*/
+
     public function create(){
 
         $accountsQuery = Animal::all();
@@ -34,7 +57,10 @@ class AnimalController extends Controller
         $animal = $this->validate(request(), [
             'name' => 'required',
             'date_of_birth' => 'required',
-            'image' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg|max:500'
+            'image' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg|max:500',
+            'image2' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg|max:500',
+            'image3' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg|max:500'
+
         ]);
         //Handles the uploading of the image
         if($request->hasFile('image')){
@@ -53,14 +79,39 @@ class AnimalController extends Controller
             $fileNameToStore = 'noimage.jpg';
         }
 
-        // create a animal object and set its values from the input
+        if($request->hasFile('image2')){
+            $fileNameWithExt = $request->file('image2')->getClientOriginalName();
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('image2')->getClientOriginalExtension();
+            $fileNameToStore2 = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('image2')->storeAs('public/images', $fileNameToStore2);
+            }
+        else{
+            $fileNameToStore2 = 'noimage.jpg';
+        }
+
+        if($request->hasFile('image3')){
+            $fileNameWithExt = $request->file('image3')->getClientOriginalName();
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('image3')->getClientOriginalExtension();
+            $fileNameToStore3 = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('image3')->storeAs('public/images', $fileNameToStore3);
+            }
+        else{
+            $fileNameToStore3 = 'noimage.jpg';
+        }
+
         $animal = new Animal;
         $animal->name = $request->input('name');
         $animal->date_of_birth = $request->input('date_of_birth');
         $animal->description = $request->input('description');
+        $animal->animal_type = $request->input('animal_type');
         $animal->availability = $request->input('availability');
         $animal->created_at = now();
         $animal->image = $fileNameToStore;
+        $animal->image2 = $fileNameToStore2;
+        $animal->image3 = $fileNameToStore3;
+
         // save the animal object
         $animal->save();
         // generate a redirect HTTP response with a success message
@@ -69,15 +120,13 @@ class AnimalController extends Controller
     }
 
 
-    public function index(){
-        $animals = Animal::all()->toArray();
-        return view('animals.index', compact('animals'));
-    }
 
     public function destroy($id){
-        $animal = Animal::find($id);
-        $animal->delete();
-        return redirect('animals');
+        if (!Gate::denies('add_animals')) {
+            $animal = Animal::find($id);
+            $animal->delete();
+            return redirect('animals');
+        }
     }
 
     public function edit($id){
@@ -91,7 +140,7 @@ class AnimalController extends Controller
         $animal = Animal::find($id);
         $this->validate(request(), [
             'name' => 'required',
-            //'date_of_birth' => 'required|numeric'
+            'date_of_birth' => 'required'
         ]);
         $animal->name = $request->input('name');
         $animal->date_of_birth = $request->input('date_of_birth');
@@ -118,9 +167,6 @@ class AnimalController extends Controller
         $animal->save();
         return redirect('animals');
     }
-
-
-
